Add files via upload
In the Zi paper they looking at images with only cardiomegaly and only edema not cardiomegaly and edema combined with other images so we should probably do that instead so we can compare better with that paper. I can do that later. Also, I complied all the images into one folder since they are in multiple folders and subfolders (this might be annoying to do if you have an older computer).